Updated error handler util
