+++ Tests.GoogleSheets/SpreadsheetTests.cs
@@ -52,7 +52,7 @@ namespace Tests.GoogleSheets
             var action = new SpreadsheetActions(InvocationContext, FileManager);
 
             var spreadsheetFileRequest = new SpreadsheetFileRequest { SpreadSheetId = "" };
-            var spreadSheet = new SheetRequest { SheetName = "YOUR_FILE_NAME" };
+            var spreadSheet = new SheetRequest { SheetName = "" };
             var rangeRequest = new OptionalRangeRequest {  };
 
             var result = await action.DownloadCSV(spreadsheetFileRequest, spreadSheet, rangeRequest, new CsvOptions { });
